modify codes for p2p communication

+++ go-seele-one-shard/seele/download/peer.go
@@ -27,9 +27,9 @@ var (
 
 // Peer define some interfaces that request peer data
 type Peer interface {
-	HeadByChain(chainNum uint) (common.Hash, *big.Int)
-	RequestHeadersByHashOrNumber(magic uint32, origin common.Hash, chainNum uint, num uint64, amount int, reverse bool) error
-	RequestBlocksByHashOrNumber(magic uint32, origin common.Hash, chainNum uint, num uint64, amount int) error
+	HeadByChain(chainNum uint64) (common.Hash, *big.Int)
+	RequestHeadersByHashOrNumber(magic uint32, origin common.Hash, chainNum uint64, num uint64, amount int, reverse bool) error
+	RequestBlocksByHashOrNumber(magic uint32, origin common.Hash, chainNum uint64, num uint64, amount int) error
 	GetPeerRequestInfo() (uint32, common.Hash, uint64, int)
 }
